deployment/requirements.txt
streamlit
pandas
tensorflow-cpu
sklearn